:recycle: [Refactor]: Gemini code assist의 코드 리뷰 내용 반영
- entities/calendar/api 세그먼트의 공개 API 수정
- UI 관련 타입 또는 해당 컴포넌트에서만 사용되는 타입을 model이 아닌 lib에 위치하도록 변경
