+++ src/entities/calendar/api/index.ts
@@ -1,1 +1,2 @@
-export { CalendarApi } from "./api";
+export { calendarApi } from "./calendarApi";
+export { participantApi } from "./participantApi";
