+++ src/features/view-daily-schedule/lib/index.ts
