+++ src/features/view-daily-schedule/lib/types.ts
