+++ src/features/view-daily-schedule/ui/index.tsx
@@ -1,7 +1,7 @@
 import type { RightSidebarViewType } from "@/shared/model";
 import { ButtonGroup, ScrollArea } from "@/shared/ui";
 import { Clock, Plus, Share2 } from "lucide-react";
-import type { ScheduleItem } from "../model";
+import type { ScheduleItem } from "../lib";
 
 interface DailyScheduleProps {
   selectedDate?: Date;
